change WishList logic replaced to App from AppRoutes
