+++ src/App.js
@@ -2,8 +2,7 @@ import React, {useEffect} from 'react';
 import {BrowserRouter} from 'react-router-dom';
 import AppRoutes from './routes/AppRoutes';
 
-import {Provider} from 'react-redux';
-import store from './redux/store';
+import {connect} from 'react-redux';
 
 import {StylesProvider, ThemeProvider} from '@material-ui/styles';
 
@@ -11,26 +10,35 @@ import styles from './App.scss';
 import theme from './mainTheme';
 import Footer from './containers/Footer/Footer';
 import AjaxUtils from './ajax';
+import { changeWishList } from './redux/actions/actions';
+import { getStorageData } from './helpers/helpers';
 
-function App () {
+function App (props) {
+  const {changeWishList} = props;
+
   useEffect(() => {
     AjaxUtils.Categories.getAllCategories();
-  }, []);
+    changeWishList(getStorageData('wishList'));
+  }, [changeWishList]);
 
   return (
     <div className={styles.App}>
-      <Provider store={store}>
-        <BrowserRouter>
-          <StylesProvider injectFirst>
-            <ThemeProvider theme={theme}>
-              <AppRoutes/>
-              <Footer/>
-            </ThemeProvider>
-          </StylesProvider>
-        </BrowserRouter>
-      </Provider>
+      <BrowserRouter>
+        <StylesProvider injectFirst>
+          <ThemeProvider theme={theme}>
+            <AppRoutes/>
+            <Footer/>
+          </ThemeProvider>
+        </StylesProvider>
+      </BrowserRouter>
     </div>
   );
 }
 
-export default App;
+const mapDispatchToProps = dispatch => {
+  return {
+    changeWishList: data => dispatch(changeWishList(data))
+  };
+};
+
+export default connect(null, mapDispatchToProps)(App);
